[Bug] Move issue-close tests to their own file
